code/requirements.txt
nni
pytorch
configobj
tensorboard
